Time Series Modeling and Intervention Analysis on Tourist Arrivals in Nepal
Predicted 2019 and 2020 monthly tourist arrivals in Nepal using a multiplicative seasonal ARIMA model with data from 1992 to 2018. Intervention analysis identified 2015 earthquake as temporary change outlier with an immediate effect of 64% reduction of arrivals in the first period and subsequent gradual decrease of impact, which was non-existent by the end of 2018.